Allow the SIP Phone Plugin to be compiled by Maven.
Note 1: The assembly process does not currently include the JainSipAPI1.1.jar
file because it's not a proper dependency. It's therefore necessary to
copy this manually in to the expanded lib folder for anything to run
Note 2: I've not actually tested this in anyway whatsoever. It does
however at least compile, now.